requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.14.2
numpy==2.1.3
openai==1.55.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
py-dotenv==0.1
pyarrow==18.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.0
safehttpx==0.1.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
websockets==12.0
Werkzeug==3.1.3
